refactor. erstatt erGyldigTelefonnummer med direkte regex-validering og gi testfiler spec-suffiks
Fjerner erGyldigTelefonnummer fra tlfUtils og bruker i stedet
NORSK_TELEFONNUMMER_REGEX og UTENLANDSK_TELEFONNUMMER_REGEX direkte
i TelefonnummerInput sin zod-schema, noe som forenkler valideringslogikken.

Omdøper testfilene fra .test til .spec i henhold til konvensjon, og
oppdaterer tlfUtils.spec.ts til å fjerne tester for den slettede funksjonen.
Dekningen videreføres i TelefonnummerInput.spec.tsx med nye tester for
norsk, utenlandsk og ugyldig input.
