+++ src/utils/tlfUtils.spec.ts
@@ -1,6 +1,5 @@
 import { describe, expect, it } from 'vitest';
 import {
-    erGyldigTelefonnummer,
     formaterNorskeTelefonnummer,
     NORSK_MOBILNUMMER_REGEX,
     NORSK_TELEFONNUMMER_REGEX,
@@ -68,72 +67,6 @@ describe('NORSK_MOBILNUMMER_REGEX', () => {
 
     it('avviser nummer med spesialtegn', () => {
         expect(NORSK_MOBILNUMMER_REGEX.test('4123456!')).toBe(false);
-    });
-});
-
-describe('erGyldigTelefonnummer', () => {
-    it('godtar norsk 8-sifret nummer uten landkode', () => {
-        expect(erGyldigTelefonnummer('21234567')).toBe(true);
-    });
-
-    it('godtar norsk mobilnummer uten landkode', () => {
-        expect(erGyldigTelefonnummer('41234567')).toBe(true);
-    });
-
-    it('godtar norsk nummer med mellomrom', () => {
-        expect(erGyldigTelefonnummer('412 34 567')).toBe(true);
-    });
-
-    it('godtar norsk nummer med +47', () => {
-        expect(erGyldigTelefonnummer('+4741234567')).toBe(true);
-    });
-
-    it('godtar norsk nummer med 0047', () => {
-        expect(erGyldigTelefonnummer('004741234567')).toBe(true);
-    });
-
-    it('avviser norsk nummer med feil lengde', () => {
-        expect(erGyldigTelefonnummer('4123456')).toBe(false);
-    });
-
-    it('avviser norsk nummer med for mange siffer', () => {
-        expect(erGyldigTelefonnummer('412345678')).toBe(false);
-    });
-
-    it('avviser norsk nummer med for mange siffer selv med +47', () => {
-        expect(erGyldigTelefonnummer('+47412345678')).toBe(false);
-    });
-
-    it('avviser norsk nummer som starter med 0', () => {
-        expect(erGyldigTelefonnummer('01234567')).toBe(false);
-    });
-
-    it('godtar utenlandsk nummer med +', () => {
-        expect(erGyldigTelefonnummer('+34636263227')).toBe(true);
-    });
-
-    it('godtar utenlandsk nummer med 00', () => {
-        expect(erGyldigTelefonnummer('004915123456789')).toBe(true);
-    });
-
-    it('avviser utenlandsk nummer uten landkodeprefiks', () => {
-        expect(erGyldigTelefonnummer('34636263227')).toBe(false);
-    });
-
-    it('avviser nummer med spesialtegn', () => {
-        expect(erGyldigTelefonnummer('+34 636!')).toBe(false);
-    });
-
-    it('avviser for langt nummer', () => {
-        expect(erGyldigTelefonnummer('+1234567890123456')).toBe(false);
-    });
-
-    it('avviser tom streng', () => {
-        expect(erGyldigTelefonnummer('')).toBe(false);
-    });
-
-    it('avviser undefined', () => {
-        expect(erGyldigTelefonnummer(undefined)).toBe(false);
     });
 });
